+++ shard/exceptions.py
@@ -2,7 +2,7 @@ from typing import List
 
 __all__ = (
     'StrategyNotImplementException', 'RequireSpecificDatabaseException', 'RequireMasterConfigException', 'NotShardingModelException',
-    'RequireShardKeyException', 'QueryExecuteFailureException',
+    'RequireShardKeyException', 'QueryExecuteFailureException', 'NotFoundPoolException',
 )
 
 
@@ -33,3 +33,7 @@ class QueryExecuteFailureException(Exception):
         self.parent_exception = exception
 
         super().__init__(*args, **kwargs)
+
+
+class NotFoundPoolException(Exception):
+    pass