+++ shard_static/services/lock_service.py
@@ -1,11 +1,16 @@
 from importlib import import_module
 from typing import Type
 
+from django.core.exceptions import ImproperlyConfigured
+
 from shard_static import config
 from shard_static.lock import BaseLockManager
 
 
 def get_lock_manager(model_name: str, database_alias: str) -> BaseLockManager:
+    if config.SHARD_SYNC_LOCK_MANAGER_CLASS is None:
+        raise ImproperlyConfigured("Lock class is not configured. Please configure lock class.")
+
     lock_key = _make_lock_key(model_name, database_alias)
     ttl = config.SHARD_SYNC_LOCK_TTL
     lock_manager_class = _import_lock_manager_class()